arch/arm64: the arm64 perf interface supports pmu
Summary:
- Support arm64 pmu api, Currently only the cycle counter function is supported.
- Using ARM64 PMU hardware capability to implement perf interface, modify all
  perf interface related code.
- Support for pmu init under smp.

+++ drivers/segger/config/SEGGER_SYSVIEW_Conf.h
@@ -73,7 +73,7 @@ extern "C"
  ****************************************************************************/
 
 unsigned int note_sysview_get_interrupt_id(void);
-unsigned int note_sysview_get_timestamp(void);
+unsigned long note_sysview_get_timestamp(void);
 
 #undef EXTERN
 #if defined(__cplusplus)
